requirements.txt
alive-progress==1.3.0
apparmor==2.13.3
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
Automat==0.8.0
btrfsutil==1.1.1
CacheControl==0.12.6
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
distlib==0.3.0
distro==1.4.0
docutils==0.14
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
isort==4.3.21
jsonpickle==1.2
lazy-object-proxy==1.4.3
LibAppArmor==2.13.3
libmsym==0.2.4
lockfile==0.12.2
mccabe==0.6.1
more-itertools==8.0.2
msgpack==0.6.2
ordered-set==3.1.1
packaging==19.2
pep517==0.6.0
progress==1.5
pwquality==1.4.2
PyHamcrest==1.9.0
pylint==2.4.4
pyparsing==2.4.6
PyQt5==5.14.0
PyQt5-sip==12.7.0
python-digitalocean==1.14.0
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.13.0
team==1.0
Twisted==19.10.0
urllib3==1.25.7
virtualenv==16.7.9
webencodings==0.5.1
wrapt==1.11.2
zipp==0.6.0
zope.interface==4.7.1
